DEV-005 creating product overview endpoint
Added @Autowired were needed
Moved excisting @Autowired to constructor per team convention
Created new testclass for non-controller tests
Changed returntype for findAll method in order to accept Product instead of dto
Extracted testUtils classes according to high cohesion principle
