+++ src/main/java/be/syntra/devshop/DevshopBack/services/ProductService.java
@@ -1,5 +1,6 @@
 package be.syntra.devshop.DevshopBack.services;
 
+import be.syntra.devshop.DevshopBack.entities.Product;
 import be.syntra.devshop.DevshopBack.models.ProductDto;
 
 import java.util.List;
@@ -9,5 +10,5 @@ public interface ProductService {
 
     ProductDto save(ProductDto productDTO);
 
-    List<ProductDto> findAll();
+    List<Product> findAll();
 }
