+++ src/main/java/be/syntra/devshop/DevshopBack/services/ProductServiceImpl.java
@@ -1,8 +1,10 @@
 package be.syntra.devshop.DevshopBack.services;
 
+import be.syntra.devshop.DevshopBack.entities.Product;
 import be.syntra.devshop.DevshopBack.models.ProductDto;
 import be.syntra.devshop.DevshopBack.repositories.ProductRepository;
 import be.syntra.devshop.DevshopBack.services.utilities.MapperUtility;
+import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
@@ -13,14 +15,15 @@ public class ProductServiceImpl implements ProductService {
     private ProductRepository productRepository;
     private MapperUtility mapperUtility;
 
+    @Autowired
     public ProductServiceImpl(ProductRepository productRepository, MapperUtility mapperUtility) {
         this.productRepository = productRepository;
         this.mapperUtility = mapperUtility;
     }
 
     @Override
-    public List<ProductDto> findAll() {
-        return mapperUtility.convertListToDtos(productRepository.findAll());
+    public List<Product> findAll() {
+        return productRepository.findAll();
     }
 
     @Override
