+++ src/test/java/be/syntra/devshop/DevshopBack/controllers/ProductControllerTest.java
@@ -1,7 +1,9 @@
 package be.syntra.devshop.DevshopBack.controllers;
 
+import be.syntra.devshop.DevshopBack.entities.Product;
 import be.syntra.devshop.DevshopBack.models.ProductDto;
 import be.syntra.devshop.DevshopBack.services.ProductServiceImpl;
+import be.syntra.devshop.DevshopBack.services.utilities.MapperUtility;
 import org.junit.jupiter.api.Test;
 import org.mockito.Mock;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -10,12 +12,11 @@ import org.springframework.boot.test.mock.mockito.MockBean;
 import org.springframework.http.MediaType;
 import org.springframework.test.web.servlet.MockMvc;
 
-import java.math.BigDecimal;
-import java.util.LinkedList;
 import java.util.List;
 
-import static be.syntra.devshop.DevshopBack.testutilities.TestUtils.asJsonString;
-import static org.assertj.core.api.Assertions.assertThat;
+import static be.syntra.devshop.DevshopBack.testutilities.GeneralUtils.asJsonString;
+import static be.syntra.devshop.DevshopBack.testutilities.ProductUtils.createProductDto;
+import static be.syntra.devshop.DevshopBack.testutilities.ProductUtils.createProductList;
 import static org.mockito.Mockito.any;
 import static org.mockito.Mockito.when;
 import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.get;
@@ -37,25 +38,13 @@ class ProductControllerTest {
     @Mock
     private ProductDto productDto;
 
-    private ProductDto createProduct() {
-        return ProductDto.builder()
-                .name("post-its")
-                .price(BigDecimal.valueOf(1.00))
-                .build();
-    }
-
-    private List<ProductDto> createProductList() {
-        List<ProductDto> products = new LinkedList<>();
-        for (int i = 0; i < 3; i++) {
-            products.add(createProduct());
-        }
-        return products;
-    }
+    @Mock
+    private MapperUtility mapperUtility;
 
     @Test
     void addProductTest() throws Exception {
         // Given
-        ProductDto newProduct = createProduct();
+        ProductDto newProduct = createProductDto();
         // When
         when(productService.save(any(ProductDto.class))).thenReturn(newProduct);
         // Then
@@ -66,20 +55,10 @@ class ProductControllerTest {
     }
 
     @Test
-    void getAllProductsTest() {
-        // Given
-        List<ProductDto> products = createProductList();
-        // When
-        when(productService.findAll()).thenReturn(products);
-        // Then
-        assertThat(productService.findAll().size()).isEqualTo(products.size());
-    }
-
-    @Test
     void allProductsEndPointTest() throws Exception {
         // Given
-        List<ProductDto> products = createProductList();
-        ProductDto singleProduct = products.get(0);
+        List<Product> products = createProductList();
+        ProductDto singleProduct = mapperUtility.convertToProductDto(products.get(0));
         // When
         when(productService.save(singleProduct)).thenReturn(singleProduct);
         when(productService.findAll()).thenReturn(products);
